+++ src/resources/elements/ckEditor/pckeditor-text.html
@@ -1,0 +1,17 @@
+<template>
+  <!-- <input
+    ref="input"
+    class="pInput ${ validationState ? 'pInput-' + validationState : '' }"
+    disabled.bind="disabled"
+    value.bind="value"
+    placeholder.to-view="placeholder"
+    autocomplete.to-view="autocomplete"
+  /> -->
+  <div ref="editorRef"></div>
+  ${validationState}
+  <template if.bind="!disabled">
+    <i class="feedback-icon fas fa-exclamation-circle" if.bind="validationState === 'warning'"></i>
+    <i class="feedback-icon fas fa-exclamation-triangle" if.bind="validationState === 'error'"></i>
+    <i class="feedback-icon fas fa-circle-notch loading" if.bind="validationState === 'validating'"></i>
+  </template>
+</template>
